Add "read_to_release" tag
